Updating

+++ java/java-analysis-impl/src/com/intellij/codeInspection/canBeFinal/CanBeFinalAnnotator.java
@@ -30,7 +30,7 @@ import java.util.*;
  */
 class CanBeFinalAnnotator extends RefGraphAnnotatorEx {
   private final RefManager myManager;
-  public static int CAN_BE_FINAL_MASK;
+  public static long CAN_BE_FINAL_MASK;
 
   public CanBeFinalAnnotator(@NotNull RefManager manager) {
     myManager = manager;
